+++ src/Unleashed/Context/Scripting.cs
@@ -27,6 +27,38 @@ static class Scripting
 		NetworkUI.Instance.GUIContextualGUID
 			.transform.parent.Find("01 Editor")
 			.gameObject.GetOrAddComponent<UIDisableIfNotServer>();
+		NetworkUI.Instance.GUIContextualGUID
+			.GetOrAddComponent<UZContextualGuid>();
+	}
+}
+sealed class UZContextualGuid : MonoBehaviour
+{
+	void OnAltClick()
+	{
+		_ = this;
+		if (!Network.IsServer)
+			return;
+		UIDialog.ShowInput(
+			$"Enter GUID for {Pointer.MyPointer.InfoObject}",
+			inputName: Pointer.MyPointer.InfoObject.GetNPO().GUID,
+
+			leftButtonText: "OK",
+			leftButtonFunc: input =>
+			{
+				if (input is not [_, ..])
+					return;
+				foreach (var npo in ManagerPhysicsObject.Instance.GrabbableNPOs)
+				if      (npo.GUID == input)
+				{
+					Chat.LogError("GUID already in use!");
+					return;
+				}
+				Pointer.MyPointer.InfoObject.GetNPO().GUID = input;
+				Pointer.MyPointer.ResetInfoObject();
+			},
+			rightButtonText: "Cancel",
+			rightButtonFunc: null
+		);
 	}
 }